gemini/requirements.txt
streamlit
google-genai
pyaudio
nest_asyncio
python-dotenv
numpy
sounddevice
simpleaudio
#streamlit run gemini.py      